removed excess files, added .sh files to run sync_pubmed.py and create_manifest.py
