+++ sync_and_process_scripts/pubmed/sync_pubmed_data.sh
@@ -7,7 +7,7 @@ source ~/miniconda3/etc/profile.d/conda.sh
 conda activate intro-sos
 
 # Execute the Python script
-python get_pubmed.py --o './output_dir/data'
+python sync_pubmed.py --o '/home/dimu6211/pl/oa_package'  --ftp_host 'ftp.ncbi.nlm.nih.gov'  --starting_pubmed_dir '/pub/pmc/oa_package'
 
 # Deactivate the Conda environment
 conda deactivate
